# TM/Task-Management/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "The API Documentation",
        "description": "This document is used to know api's status and run the api's",
        "contact": {
            "email": "arunbalajiselvan@gmail.com"
        },
        "version": "3.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8004/public/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login User",
                "description": "Login User",
                "operationId": "postLogin",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "test@domain.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "Password@12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register User",
                "description": "Register User",
                "operationId": "postRegister",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Test User"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "test@domain.com"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "Developer"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "Password@12345"
                                    },
                                    "confirm_password": {
                                        "type": "string",
                                        "example": "Password@12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/project/store": {
            "post": {
                "tags": [
                    "Project"
                ],
                "summary": "Create project",
                "description": "create new project",
                "operationId": "createProject",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "NotifyNow"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/project/update": {
            "post": {
                "tags": [
                    "Project"
                ],
                "summary": "Update project",
                "description": "Update project",
                "operationId": "updateProject",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "PHP"
                                    },
                                    "id": {
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/project/delete/{id}": {
            "get": {
                "tags": [
                    "Project"
                ],
                "summary": "Delete project",
                "description": "Delete project by id",
                "operationId": "deleteProject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "{Need to pass id as parameter}",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "Authorization": {
                "type": "http",
                "description": "Provide token to access api",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "Authorization": []
        }
    ]
}